WebAuthentication/requirements.txt
Authlib==1.6.5
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
Flask-HTTPAuth==4.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.0
Werkzeug==3.1.4
